+++ typeClassDefs/WalkData.h
@@ -33,7 +33,7 @@ public:
     void from_string(const std::string)  {
 		for(int i = 0; i<3; ++i) {
 			for(int j = 0; j<4; ++j) {
-				data[j][i] = i*j;
+				data[j][i] = static_cast<float>(i*j);
 			}
 		}
 	}
